Fix bug due to loss of precision in ratio calculation of FP Growth. Fix bug in MinCovDet due to non-update-in-place vector semantics.

+++ src/main/java/macrobase/analysis/summary/itemset/FPGrowthEmerging.java
@@ -102,7 +102,7 @@ public class FPGrowthEmerging {
                 Integer inlierCount = inlierCounts.get(i.getItems().iterator().next());
 
                 if(inlierCount != null && inlierCount > 0) {
-                    ratio = (i.getCount()/outliers.size())/((double)inlierCount/inliers.size());
+                    ratio = ((double)i.getCount()/outliers.size())/((double)inlierCount/inliers.size());
                 } else {
                     ratio = Double.POSITIVE_INFINITY;
                 }
